Add idunion user api draft
Adds the openapi spec of the idunion user service api. It allows to
create new users, login and to get tokens. The API is still in a
draft state and likely to change.

The API gets added to the idunion-web frontend to let the HTMX server
side forward the API requests to the actual API handling service. With
that, it is not required to expose the API service directly to the
internet which keeps the amount of exposed endpoints and APIs at a
minimum.