Add user login
The user login form makes some simple validation in the frontend of the
typed in values. It shows success and error messages from the backend.

* Add mail domain as static variable

Mail domain can be configured and gets added to the static template
rendering.

Since we want to have each user belonging to idunion, the mail domain
cannot be arbitrary chosen.

* Request auth code

Requests the auth code with the credentials that got typed in by the
user

* Request access token

Uses the requested auth code to request a token from the auth server.
Actually, the redirect_uri of the auth code request shall point to the
client so that the client can use the auth code to obtain an access
token.

For testing, the backend uses the auth code imediately to obtain a token
too.

* Store access token as secure http only cookie

Stores the opaque access token (which contains no useful information) in
a secure http only same site cookie in the browser to identify the
user-agent with each request.