Request access token
Uses the requested auth code to request a token from the auth server.
Actually, the redirect_uri of the auth code request shall point to the
client so that the client can use the auth code to obtain an access
token.

For testing, the backend uses the auth code imediately to obtain a token
too.